moved important bits into package for importability
